moved files into surveyutils
