+++ testutils/cli/testutil.go
@@ -7,7 +7,7 @@ import (
 	"github.com/hinshun/vt10x"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/solo-io/go-utils/cliutils"
+	"github.com/solo-io/go-utils/surveyutils"
 	"gopkg.in/AlecAivazis/survey.v1/terminal"
 )
 
@@ -19,7 +19,7 @@ func ExpectInteractive(userinput func(*Console), testcli func()) {
 	c, state, err := vt10x.NewVT10XConsole()
 	Expect(err).NotTo(HaveOccurred())
 	defer c.Close()
-	cliutils.UseStdio(Stdio(c))
+	surveyutils.UseStdio(Stdio(c))
 	// Dump the terminal's screen.
 	defer func() { GinkgoWriter.Write([]byte(expect.StripTrailingEmptyLines(state.String()))) }()
